// frontend/src/gen/TwoFactorSetupResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { TwoFactorSetupState } from "./TwoFactorSetupState";

export type TwoFactorSetupResponse = { "kind": "ready", setup: TwoFactorSetupState, } | { "kind": "recoveryCodes", codes: Array<string>, signedOut: number, continueUrl: string, } | { "kind": "navigate", location: string, } | { "kind": "error", message: string, setup: TwoFactorSetupState, };